Remove --static flag to fix build on Arch
The final binary was not statically linked anyway so there really was not a
strong reason to prefer static linking. Removing this fixes an issue when trying
to link against the Arch capstone package which only includes the .so. capstone
is used for pretty-printing disassembly info on MTE tag check failures so I'd
rather not disable the lib.